backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
psutil==6.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.6.0
apscheduler==3.10.4
httpx==0.28.0
paramiko==3.5.0
docker==7.1.0
pytest==8.3.0
pytest-asyncio==0.24.0
